\name{ensembleGGM-package}
\alias{ensembleGGM}
\title{ensembleGGM: ensemble estimation for Gaussian graphical models}
\description{
This package estimates a precision matrix (and thereby a GGM) using the SpiderLearner algorithm of Shutta et al. (2021).

The package is designed in an object-oriented framework, where a SpiderLearner object contains a library of Candidate objects, each of which has a member function fit() that estimates a precision matrix. See ?MakeSpiderLearner for a minimum working example of how to use these tools.

}
\author{
Kate Hoff Shutta (kshutta@hsph.harvard.edu)
}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
